API

API

Model

ExampleModeling.ExampleModel — Type.

Defines the ExampleModel type.

source

ExampleModeling.Specs — Type.

Specifies the control flow of the model. For example, it can be used to toggle on and off specific constraints or objectives.

source

ExampleModeling.Indices — Type.

Contains indices of the model.

source

ExampleModeling.Params — Type.

Contains the parameters of the model.

source

ExampleModeling.Variables — Type.

Contains the variable values of the model.

source

ExampleModeling.Objectives — Type.

Contains the objective values of the model.

source

ExampleModeling.ExampleModel — Method.

Initializes the ExampleModel.

source

ExampleModeling.Variables — Method.

Queries the variable values from the model into Variables type.

source

ExampleModeling.Objectives — Method.

Queries the objective values from the model into Objectives type.

source

IO

ExampleModeling.save_json — Function.

Saves the given DataType into JSON file.

Arguments

dtype

output_path::AbstractString: Full filepath, e.g., path.json.

source

ExampleModeling.load_json — Function.

Loads values from JSON file to given DataType.

Arguments

dtype: DataType

filepath::AbstractString

source

Utility

ExampleModeling.model_to_dtype — Function.

The function queries values from the model to data type based on its field names. It extracts values from DenseAxisArray from its data field. Then, it converts the values to the corresponding field type. The function rounds integers before conversion because JuMP outputs integer variables as floats.

Arguments

dtype: DataType.

model::Model: JuMP.Model.

source