API

API

Model

ExampleModeling.ExampleModel — Type.

Defines the ExampleModel type.

source

ExampleModeling.Specs — Type.

Specification for different model scenarios. For example, we can specify toggling on and off certain constraints and objectives.

Arguments

relax_integer::Bool=false: If true, relax integer constraints from variables.

source

ExampleModeling.Indices — Type.

Contains indices of the model.

source

ExampleModeling.Params — Type.

Contains the parameters of the model.

source

ExampleModeling.Variables — Type.

Contains the variable values of the model.

source

ExampleModeling.Objectives — Type.

Contains the objective values of the model.

source

ExampleModeling.ExampleModel — Method.

Initializes the ExampleModel.

source

ExampleModeling.Variables — Method.

Queries the variable values from the model into Variables type.

source

ExampleModeling.Objectives — Method.

Queries the objective values from the model into Objectives type.

source

IO

ExampleModeling.save_json — Function.

Saves any JSON serializable object into JSON file in filepath.

source

ExampleModeling.load_json — Function.

Loads values from JSON file in filepath to DataType supplied by dtype.

source

Utility

ExampleModeling.model_to_dtype — Function.

The function queries values from the model to data type based on its field names. It extracts values from DenseAxisArray from its data field. Then, it converts the values to the corresponding field type. The function rounds integers before conversion because JuMP outputs integer variables as floats.

source