API

Model

ExampleModeling.ExampleModel — Type

Defines the ExampleModel type.

ExampleModeling.Specs — Type

Specification for different model scenarios. For example, we can specify toggling on and off certain constraints and objectives.

Arguments

relax_integer::Bool=false: If true, relax integer constraints from variables.

ExampleModeling.Indices — Type

Contains indices of the model.

ExampleModeling.Params — Type

Contains the parameters of the model.

ExampleModeling.Variables — Type

Contains the variable values of the model.

ExampleModeling.Objectives — Type

Contains the objective values of the model.

ExampleModeling.ExampleModel — Method

Initializes the ExampleModel.

ExampleModeling.Variables — Method

Queries the variable values from the model into Variables type.

ExampleModeling.Objectives — Method

Queries the objective values from the model into Objectives type.

IO

ExampleModeling.save_json — Function

Saves any JSON serializable object into JSON file in filepath.

ExampleModeling.load_json — Function

Loads values from JSON file in filepath to DataType supplied by dtype.

Utility

ExampleModeling.model_to_dtype — Function

The function queries values from the model to data type based on its field names. It extracts values from DenseAxisArray from its data field. Then, it converts the values to the corresponding field type. The function rounds integers before conversion because JuMP outputs integer variables as floats.